More "complete" cleanup
Call setup.py clean --all to clean up more build files, which is
generally what people expect.

Also clean up .egg-info because it trips people up sometimes. And the
build env as well. We can add flags to conditionally keep them in the
future if we need to, but let's be as strict as possible now.